Naukariportal/requirements.txt
ak-djangorestframework-jsonapi==2.4.0.post0
asgiref==3.5.2
beautifulsoup4==4.11.1
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==4.1.2
django-extensions==3.2.1
django-import-export==3.0.1
django-location-field==2.1.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.8.0
Flask==2.2.2
greenlet==2.0.1
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.31
mysqlclient==2.1.1
numpy==1.23.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.5.0
pathspec==0.10.1
phonenumbers==8.13.0
phonenumberslite==8.13.0
Pillow==9.3.0
platformdirs==2.5.2
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
render==1.0.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
sqlparse==0.4.3
tablib==3.2.1
tomli==2.0.1
tweepy==4.12.1
typing_extensions==4.4.0
tzdata==2022.4
urllib3==1.26.13
views==0.3
virtualenv==20.16.5
Werkzeug==2.2.2
xlrd==2.0.1
xlwt==1.3.0
zipp==3.10.0
